+++ src/lib/CustomSystem.cc
@@ -13,32 +13,16 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-#include <es/GlobalSystem.h>
+#include <es/CustomSystem.h>
 
 namespace es {
 
-  bool GlobalSystem::addEntity(Entity e) {
-    auto ret = m_entities.insert(e);
-    return ret.second;
+  bool CustomSystem::addEntity(Entity e) {
+    return false;
   }
 
-  bool GlobalSystem::removeEntity(Entity e) {
-    auto ret = m_entities.erase(e);
-    return ret > 0;
-  }
-
-  void GlobalSystem::update(float delta) {
-    /* make a copy so that the entities can be safely removed from the system
-     * without invalidating the iterators.
-     */
-    auto copy = getEntities();
-    for (Entity e : copy) {
-      updateEntity(delta, e);
-    }
-  }
-
-  void GlobalSystem::updateEntity(float delta, Entity entity) {
-    // nothing by default
+  bool CustomSystem::removeEntity(Entity e) {
+    return false;
   }
 
 }